requirements.txt
NumPy >= 1.18.5
Pillow >= 7.2.0